reorg folders and refactor header bar
