+++ src/components/Cards/CardFocusBtn.jsx
@@ -1,10 +1,10 @@
 import React, { useState, useRef, useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { useSelector } from "react-redux"
-import { GetTodoList } from "../selectors";
+import { GetTodoList } from "../../selectors";
 import { IconButton } from '@mui/material';
 import ParaglidingIcon from '@mui/icons-material/Paragliding';
-import { isLocationExist } from '../utils/generalUtils';
+import { isLocationExist } from '../../utils/generalUtils';
 
 export const CardFocusBtn = ({id}) => {
