+++ src/components/Cards/CardMapBtns.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useRef, useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { useSelector } from "react-redux"
-import { GetTodoList } from "../selectors";
+import { GetTodoList } from "../../selectors";
 import { CardPinBtn } from './CardPinBtns';
 import { CardFocusBtn } from './CardFocusBtn';
