+++ src/components/Cards/CardPinBtns.jsx
@@ -1,9 +1,9 @@
 import React, { useState, useRef, useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { useSelector } from "react-redux"
-import { GetMapPoints, GetTodoList } from "../selectors";
+import { GetMapPoints, GetTodoList } from "../../selectors";
 import { IconButton } from '@mui/material';
-import { activeMapPinTODOMode, editTODO, editAllTODOS} from '../actions/actions';
+import { activeMapPinTODOMode, editTODO, editAllTODOS} from '../../actions/actions';
 import PushPinIcon from '@mui/icons-material/PushPin';
 import SaveIcon from '@mui/icons-material/Save';
 import ClearIcon from '@mui/icons-material/Clear';
