+++ src/reducers/uiReducer/mainInputReducer/inputValue.js
@@ -1,0 +1,11 @@
+const initialState  = ''
+
+export const inputValue = ( state = initialState , action) => {
+
+    switch(action.type) {
+        case 'handleInputType':
+            return action.updateInputValue    
+        default:
+            return state
+    }
+}
