+++ src/reducers/uiReducer/mainInputReducer/isEmpty.js
@@ -1,0 +1,12 @@
+import { inputValue } from "./inputValue"
+
+const initialState  = true
+
+export const isEmpty = () => {
+    if (inputValue !== '') {
+        return false
+    }
+    else{
+        return initialState
+    }
+}
