+++ src/selectors.jsx
@@ -7,4 +7,8 @@ export const GetMapMode = state => state.UI.Map.mode
 
 export const GetMapPoints = state => state.UI.Map.points
 
+export const GetMainInputValue = state => state.UI.mainInput.inputValue
+
+export const GetMainInputIsEmpty = state => state.UI.mainInput.isEmpty
+
